Big Bazaar Sales
Big Bazaar have collected 2013 sales data for 1559 products across 10 stores in different cities. The aim is to build a predictive model and find out the sales of each product at a particular store. Using this model, Big Bazaar will try to understand the properties of products and stores which play a key role in increasing sales.